% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_lgi.R
\name{calculate_average_lgi_stats}
\alias{calculate_average_lgi_stats}
\title{calculate_average_lgi_stats}
\usage{
calculate_average_lgi_stats(x, lgi_roi_list = NULL)
}
\arguments{
\item{x}{a dataframe output by \code{\link{readstats_subject}}}

\item{lgi_roi_list}{a named list of lists containing paired vectors, e.g.:

\if{html}{\out{<div class="sourceCode">}}\preformatted{ list(
   lh_cortex_lgi = list(values = c("lh_roi1_lgi", "lh_roi2_lgi", ...),
                       weights = c("lh_roi1_area", "lg_roi2_area", ...),
   rh_cortex_lgi = list(values = c("rh_roi1_lgi", "rh_roi2_lgi", ...),
                       weights = c("rh_roi1_area", "rh_roi2_area", ...),
 )
}\if{html}{\out{</div>}}

Each vector must contain column names from \code{x}.
The default (\code{NULL}) attempts to determine these automatically by simple
matching of prefix and suffix using typical naming rules.
The names the list elements are used as the new variable names for the
calculated weighted averages.}
}
\value{
A data.frame with rows corresponding to \code{x} where columns contain the
weighted averages specified in \code{lgi_roi_list}. Default: lh_cortex_lgi
and rh_cortex_lgi.
}
\description{
Calculates a per-hemisphere weighted average of lgi.
Unless \code{lgi_roi_list} is specified, this function expects a
column naming system consistent with output from
\code{\link{readstats_subject}}.
The default behaviour is to weight LGI by cortical surface area, but
the lgi_roi_list structure can be used to calculate generic weighted
averages within x.
}
