% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_df_to_qdec}
\alias{convert_df_to_qdec}
\title{convert_df_to_qdec}
\usage{
convert_df_to_qdec(data, fsvar = "fsid", outroot = NA,
  factor.levels = T)
}
\arguments{
\item{data}{object coercible to \code{\link{data.frame}}}

\item{fsvar}{The variable in \code{data} containing the freesurfer id codes
(cooercible to case-correct character)}

\item{outroot}{path + root to output files: \itemize{
\item{\code{outroot.table.dat} - table directly input to QDEC}
\item{\code{factor.levels} - factor level labels (one per factor)}
}
If missing (NA) then nothing is written.}

\item{factor.levels}{Should factor.levels label files be written.}
}
\value{
a data.frame for inspection
    (function is primarily used for it's side effect of writing out files)
}
\description{
This function converts a data.frame with an indexing variable of freesurfer
    subject IDs (\code{fsvar}) into a format ready for qdec.
    It will optionally write out
}
\examples{
df <- data.frame(fsid = c("s1", "s2", "s3"),
    A = 1:3,
    B = letters[1:3],
    C = as.factor(letters[4:6]),
    D = as.ordered(letters[7:9]))
convert_df_to_qdec(df, fsvar = "fsid")
\dontrun{
convert_df_to_qdec(df, fsvar = "fsid", outroot = "example_qdec_table")
}
}
