% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fmt_atlaslabel}
\alias{fmt_atlaslabel}
\title{fmt_atlaslabel}
\usage{
fmt_atlaslabel(x, remove_hemi = FALSE, remove_measure = TRUE, shorten = TRUE)
}
\arguments{
\item{x}{object coercible to character}

\item{remove_hemi}{strip hemisphere from start ^lh_|^rh_}

\item{remove_measure}{strip measure from end _lgi$|_vol$|_thickness$|_area$}

\item{shorten}{make shorter labels via common anatomical abbreviations
e.g. amyg = amygdala, OFC = orbitofrontal, medial = m}
}
\value{
a character vector
}
\description{
Common manipulations for atlas labels: remove hemisphere,
remove measure, shorten labels.
}
\details{
This is a WIP developed using the DKT atlas. Note: this will not check if
the result has as many unique levels as the input. If using this function,
(particularly with the shorten option) verify it hasn't accidentally
collapsed levels.
}
