% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fsprettify}
\alias{fsprettify}
\title{fsprettify}
\usage{
fsprettify(x, remove_hemi = FALSE, remove_measure = TRUE, shorten = TRUE)
}
\arguments{
\item{x}{object coercible to character}

\item{remove_hemi}{strip hemisphere infomation ^lh_|^rh_}

\item{remove_measure}{strip measure infomation _lgi$|_vol$|_thickness$|_area$}

\item{shorten}{make common anatomical substitutions
e.g. amyg = amygdala, OFC = orbitofrontal, medial = m}
}
\value{
a prettified character vector
}
\description{
Attempts to make a character vector of region names short and pretty.
}
\details{
This is a WIP developed using the DKT atlas.
}
