% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fsroi_info}
\alias{fsroi_info}
\title{fsroi_info}
\format{
A data.frame with 200 rows and 19 variables:
\describe{
\item{Hemi}{Left (lh) or Right (rh) hemisphere}
\item{ROI}{standardised ROI name (lower case, alphabetical)}
\item{sROI}{abbreviated standardised ROI name)}
\item{Measure}{volume, cortical thickness,
surface area, local gyrification index}
\item{lobecing}{A coarse classification of ROIs into lobes
(and cingulate cortex)}
\item{lobe}{A coarse classification of ROIs into lobes}
\item{Intensity}{Voxel value in FreesurferColorLUT.txt}
\item{cog?}{Center of Gravity for ROI (volume based)}
\item{col?}{FreesurferColorLUT.txt colour (R, G, B, Alpha)}
}
}
\usage{
fsroi_info
}
\description{
a dataset of meta-information about common Freesurfer regions/measures
}
\keyword{datasets}
