% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_norms.R
\name{norm_a_region}
\alias{norm_a_region}
\title{norm_a_region}
\usage{
norm_a_region(
  value,
  region,
  participant_age,
  gender,
  eTIV,
  magnetic_field_strength = 0L,
  modality_manuf_id,
  normstab = dkt_norms
)
}
\arguments{
\item{value}{The value of an observed quantity (e.g. thickness, surfacearea)
in a particular region.}

\item{region}{The name of the region in which the quantity was observed.
(expects names from readstats_subject* functions,
i.e. in \code{normstab$region}).}

\item{participant_age}{The age in years of the participant.
Can be decimals (e.g. 21.6)}

\item{gender}{Dummy coded. Either 1L (for Male) or 0L (for Female).}

\item{eTIV}{Freesurfer's eTIV measurement in mm3 units (e.g. 1415670)}

\item{magnetic_field_strength}{Dummy coded.
Either 1L (for 1.5T), or 0L (for 3T).}

\item{modality_manuf_id}{Name of MRI scanner's manufacturer. One of
\code{c("Siemens","Phillips", "GE")}.}

\item{normstab}{A data.frame of regression coefficients and operations in
a particular expected format. By default the internal data.frame
\code{dkt_norms} is used. This was adapted from the DKT.csv file
supplied in the supplementary material of Potvin et al 2017.
This only works for the cortical DKT parcellation and aseg regions.}
}
\value{
a single numeric value
}
\description{
Produce a Z-score for a single freesurfer parcellation measure in a single
subject. Norms are based on Potvin et al 2017.
}
