% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_norms.R
\name{norm_statsdf}
\alias{norm_statsdf}
\title{norm_statsdf}
\usage{
norm_statsdf(
  statsdf,
  etiv_colname = "etiv",
  age_colname = "age",
  gender_colname = "gender",
  id_colname = "fsid",
  mfs_colname = "fieldstrengthcode",
  manufid_colname = "mfrcode",
  normstab = dkt_norms
)
}
\arguments{
\item{statsdf}{a data.frame (probably derived from
\code{\link{readstats_subjectlist}}).}

\item{etiv_colname}{The column name of the variable containing estimated
Total Intracranial Volume (mm^3).}

\item{age_colname}{The column name of the variable containing participants
age in years (must be numeric, can be decimal).}

\item{gender_colname}{The column name of the variable containing participants
gender, coded as 0 for Female, and 1 for Male.}

\item{id_colname}{The column name containing the freesurfer ID.}

\item{mfs_colname}{The column name of the variable containing the magnetic
field strength codes. Either 0L (3T) \emph{or} 1L (1.5T).}

\item{manufid_colname}{The column name of the variable containing the
MRI scanner's manufacturer. Values must be one of
\code{c("Siemens","Phillips", "GE")}.}

\item{normstab}{A data.frame of regression coefficients and operations in
a particular expected format. By default the internal data.frame
\code{dkt_norms} is used. This was adapted from the DKT.csv file
supplied in the supplementary material of Potvin et al 2017.
This only works for the cortical DKT parcellation and aseg regions.}
}
\value{
A data.frame of z-scored data. Variable names from \code{statsdf}
are prefixed by \code{z_}.
}
\description{
Produce Z-scores based on Potvin et al 2017's norming procedure for \emph{all}
freesurfer parcellation measures in a data.frame.
See \code{\link{norm_a_region}}
for key arguments. This function infers \code{value} and
\code{region} from the data.frame column names and contents. It is
designed to work with the output of \code{\link{readstats_subjectlist}}
to which columns containing values for the following must be added:
\itemize{
\item{Age (in years)}
\item{Gender (0|1; 1 codes Male)}
\item{Scanner strength (0|1; 0 codes 3T, 1 codes 1.5T)}
\item{Scanner manufacturer ("GE", "Philips", or, "Siemens")}
}
Note that the eTIV (in mm^3; generated by
\code{\link{readstats_subjectlist}}) is also used in the norming process.
}
\section{Missing Data}{
 The algorithm cannot handle missing data. The
most defensible approach is not to calculate norms if the covariates
are unknown. Alternatively, multiple imputation with relevant
auxiliary variables (see \code{\link[mice]{mice}} could be used.
For quick approximation the following approaches might
help: First, for unknown age a reasonable guess such as study
average could be used.
For missing categorical data (Gender, scanner info)
norms could be calculated for each potential combination
of the missing values, then either average the z-scores, or use the
combination that produces the smallest z-scores (i.e. minimum norm).
There are currently no functions in this package to handle missing data
or collate results over imputations.
}

\section{Age Range}{

The normative data came from male and female subjects aged 18-94,
do not expect usable results for participants much under 18 (or over 94)
as these are extrapolated outside the observed range.
}

