% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc_etiv}
\alias{qc_etiv}
\title{qc_eTIV}
\usage{
qc_etiv(x, etiv_lab = "etiv")
}
\arguments{
\item{x}{A data.frame of Freesurfer stats with rownames set to fsid.
(see \code{\link{readstats_subject}} and \code{\link{qc_prep}}).}

\item{etiv_lab}{The quoted column name of eTIV data within \emph{x}}
}
\value{
A data.frame containing studentised residuals, adjusted, and
unadjusted p.values (sorted by p).
}
\description{
Run simple outlier detection on a single variable (designed for eTIV).
Copies the basic logic from the \code{car} package's
\code{\link[car]{outlierTest}} function, but applies adjustment for the
False Discovery Rate using the B-H procedure instead of Bonferroni.
The top 6 most significant results are printed to console.
}
\section{Interpretation}{

I suggest carefully inspecting Talairach registrations for any subjects
who meet \code{p_adj < 0.05}.
Even when no subjects meet this threshold it is a good idea to note and
perhaps inspect the subjects with the largest and smallest values of
\code{rstudent} to check if results are reasonable.
}

