% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc_prep}
\alias{qc_prep}
\title{qc_prep}
\usage{
qc_prep(x, trim = 0.2)
}
\arguments{
\item{x}{A data.frame from \code{\link{readstats_subjectlist}}}

\item{trim}{Remove this fraction of the data before calculating means/sds.
Note that \code{0.5 * trim} is removed from each tail.}
}
\description{
Prepare a Freesurfer dataset (output by \code{\link{readstats_subjectlist}})
for QC analysis.
This function:
\itemize{
\item Removes \code{fsid} from the data.frame and uses it for rownames
(leaving numeric only variables)
\item Robustly scales (with \code{\link[base]{scale}}) all remaining variables
using the trimmed mean and sd.
}
}
