% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{qc_rstudent}
\alias{qc_rstudent}
\title{qc_rstudent}
\usage{
qc_rstudent(x, etiv_lab = "etiv")
}
\arguments{
\item{x}{A data.frame of Freesurfer measures with fsid in the rownames.}

\item{etiv_lab}{The variable name for the eTIV control variable in \code{x}.
All other variables in \code{x} will be adjusted for linear effects of
\code{etiv_lab}.}
}
\value{
A data.frame containing studentised residuals
(see: \code{\link[stats]{rstudent}}).
}
\description{
Obtain studentised residuals from a linear model of measure ~ eTIV
with a robust M-estimator fit
(employs \code{\link[MASS]{rlm}} with default settings).
}
