% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_stats.R
\name{readstats_subjectlist}
\alias{readstats_subjectlist}
\title{readstats_subjectlist}
\usage{
readstats_subjectlist(
  s = list_fs_subjects(),
  dir_path = NULL,
  SUBJECTS_DIR = get_SUBJECTS_DIR(),
  atlas = c("DKT", "Desikan", "Destrieux", "Custom"),
  hemi = c("both", "lh", "rh"),
  lgi = TRUE,
  read_aseg = TRUE,
  custom_atlas_name = NULL
)
}
\arguments{
\item{s}{a vector of subject IDs for subjects in \code{SUBJECTS_DIR}
(alternatively labels for subjects if using \code{dir_path}).
Default \code{s} will use \code{\link{list_fs_subjects}} to read all
valid freesurfer subjects from \code{SUBJECTS_DIR} environment variable.}

\item{dir_path}{a vector of full paths to freesurfer directories}

\item{SUBJECTS_DIR}{Specify the Freesurfer SUBJECTS_DIR directory.
SUBJECTS_DIR is ignored if \code{dir_path} is set.}

\item{atlas}{Which cortical parcellation atlas to read?
Three built-in atlases, or "Custom" to specify using the
\code{custom_atlas_name} argument.}

\item{hemi}{Which hemisphere(s) to export stats from.}

\item{lgi}{bool. TRUE: look for additional lgi files for the specified atlas.}

\item{read_aseg}{bool. Read aseg stats as well as cortical data?}

\item{custom_atlas_name}{If the cortical \code{atlas} = "Custom", then
\code{custom_atlas_name} should contain the string that identifies the
associated stats file after the hemisphere indicator: "?h."
For example, \code{"BA_exvivo"} would import lh.BA_exvivo.stats and
rh.BA_exvivo.stats and if lgi = TRUE the ?h.BA_exvivo.pial_lgi.stats
files.}
}
\value{
Given \emph{n} subject IDs returns a \emph{n} row data.frame containing stats
values. Variable names follow hemi_region_measure format unless
whole-brain measures. \code{bl_} indicates a region spanning left and
right hemispheres.
}
\description{
Apply \code{\link{readstats_subject}} to \strong{lots} of subject IDs and collate
into a data.frame.
Has two modes of operation: \code{SUBJECTS_DIR} and \code{dir_path}.
\code{dir_path} is a list of direct paths to freesurfer subject folders, it is
paired with the vector \code{s} which is used to label the subjects in
\code{dir_path}.
\code{SUBJECTS_DIR} is a directory containing all your subject folders, if
this is set then \code{s} is a vector of subjects that exist within this
directory.
\emph{Use \code{dir_path} if you don't have a typical freesurfer setup.
If all subjects are in a single directory use \code{SUBJECTS_DIR}.}
}
\examples{
\dontrun{
## SUBJECTS_DIR mode:
readstats_subjectlist(s = c("s1", "s2", "s3"),
SUBJECTS_DIR = "/path/to/SUBJECTS")
## dir_path mode:
readstats_subjectlist(s = c("s1", "s2", "s3"),
dir_path = c("path/to/s1", "path/to/s2", "path/to/s3"))
}
}
\seealso{
readstats_subject
}
