% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fmt_atlaslabel}
\alias{fmt_atlaslabel}
\title{fmt_atlaslabel}
\usage{
fmt_atlaslabel(x, remove_hemi = FALSE, remove_measure = TRUE, shorten = TRUE)
}
\arguments{
\item{x}{object coercible to character}

\item{remove_hemi}{strip hemisphere from start ^lh_|^rh_}

\item{remove_measure}{strip measure from end _lgi$|_vol$|_thickness$|_area$}

\item{shorten}{make shorter labels via common anatomical abbreviations
e.g. amyg = amygdala, OFC = orbitofrontal, medial = m}
}
\value{
a character vector
}
\description{
Common manipulations for freesurfer atlas labels: forces lower case then
optionally removes hemisphere, removes measure and shortens labels with
common abbreviations.
}
\details{
This is a WIP developed/tested with the DKT atlas.

This function does not check that the result has as many unique levels as the
input (sometimes this is the desired behaviour). When using this function,
(particularly with the shorten option) verify it hasn't collapsed levels
in a way you were not expecting.
}
