% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_norms.R
\name{norm_statsdf}
\alias{norm_statsdf}
\title{norm_statsdf}
\usage{
norm_statsdf(
  statsdf,
  etiv_colname = "etiv",
  age_colname = "age",
  gender_colname = "gender",
  id_colname = "fsid",
  magnetic_field_strength = 0L,
  modality_manuf_id,
  normstab = dkt_norms
)
}
\arguments{
\item{statsdf}{a data.frame (probably derived from
\code{\link{readstats_subjectlist}}).}

\item{etiv_colname}{The column name of the variable containing estimated
Total Intracranial Volume (mm^3).}

\item{age_colname}{The column name of the variable containing participants
age in years (must be numeric, can be decimal).}

\item{gender_colname}{The column name of the variable containing participants
gender, coded as 0 for Female, and 1 for Male.}

\item{id_colname}{The column name#' containing the freesurfer ID.}

\item{magnetic_field_strength}{Dummy coded.
Either 1L (for 1.5T), or 0L (for 3T).}

\item{modality_manuf_id}{Name of MRI scanner's manufacturer. One of
\code{c("Siemens","Phillips", "GE")}.}

\item{normstab}{A data.frame of regression coefficients and operations in
a particular expected format. By default the internal data.frame
\code{dkt_norms} is used. This was adapted from the DKT.csv file
supplied in the supplementary material of Potvin et al 2017.
This only works for the cortical DKT parcellation and aseg regions.}
}
\value{
A data.frame of z-scored data. Variable names from \code{statsdf}
are prefixed by \code{z_}.
}
\description{
Produce Z-scores based on Potvin et al 2017's norming procedure for \emph{all}
freesurfer parcellation measures in a data.frame.
See \code{\link{norm_a_region}}
for key arguments. This function infers \code{value} and
\code{region} from the data.frame column names and contents. It is
designed to work with the output of \code{\link{readstats_subjectlist}}.
}
\details{
This function assumes \code{statsdf} contains data from a single scanner,
and so \code{magnetic_field_strength} and \code{modality_manuf_id} are
specified once. Other non-region/value data required for norming
must be included as columns in \code{statsdf}.
These are: \code{participant_age} (in years),
\code{gender} (0|1 = Female|Male),
and \code{eTIV} (in mm^3).
}
