% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_stats.R
\name{readstats_subjectlist}
\alias{readstats_subjectlist}
\title{readstats_subjectlist}
\usage{
readstats_subjectlist(
  s,
  dir_path = NULL,
  SUBJECTS_DIR = NULL,
  atlas = c("DKT", "Desikan", "Destrieux"),
  hemi = c("both", "lh", "rh"),
  lgi = TRUE,
  read_aseg = TRUE
)
}
\arguments{
\item{s}{a vector of subject IDs (or labels if using dir_path)}

\item{dir_path}{Specify a path to a particular Freesurfer stats folder
(e.g. "/data/study/fs/S01/stats")}

\item{SUBJECTS_DIR}{Specify the Freesurfer SUBJECTS_DIR directory.
If \code{NULL} R will try to read the environmental variable.
SUBJECTS_DIR is ignored if \code{dir_path} is set.}

\item{atlas}{Which cortical parcellation atlas to read.}

\item{hemi}{Which hemisphere(s) to export stats from.}

\item{lgi}{bool. TRUE: look for additional lgi files for the specified atlas.}

\item{read_aseg}{bool. Read aseg stats as well as cortical data?}
}
\value{
Given \emph{n} subject IDs returns a \emph{n} row data.frame containing stats
values. Variable names follow hemi_region_measure format unless
whole-brain measures. \code{bl_} indicates a region spanning left and
right hemispheres.
}
\description{
Apply \code{\link{readstats_subject}} to a vector of subject IDs and collate
into a data.frame.
}
\seealso{
readstats_subject
}
